Remove StartGameEvent, add LivingUpdateEvent, and organise events into different files.
