+++ src/main/java/net/weavemc/loader/api/command/CommandBus.java
@@ -1,8 +1,8 @@
 package net.weavemc.loader.api.command;
 
-import net.weavemc.loader.api.event.ChatSentEvent;
 import net.weavemc.loader.api.event.EventBus;
 import lombok.experimental.UtilityClass;
+import net.weavemc.loader.api.event.client.ChatEvent;
 
 import java.util.ArrayList;
 import java.util.Arrays;
@@ -18,13 +18,13 @@ public class CommandBus {
     }
 
     static {
-        EventBus.subscribe(ChatSentEvent.class, e -> {
-            if(!e.getMessage().startsWith("/")) return;
+        EventBus.subscribe(ChatEvent.Sent.class, e -> {
+            if (!e.getMessage().startsWith("/")) return;
 
             String[] split = e.getMessage().substring(1).split("\\s+");
 
             Iterator<Command> matching = commands.stream().filter(c -> c.matches(split[0])).iterator();
-            if(!matching.hasNext()) return;
+            if (!matching.hasNext()) return;
 
             e.setCancelled(true);
             String[] args = Arrays.copyOfRange(split, 1, split.length);
