+++ src/main/kotlin/net/weavemc/loader/api/event/Event.kt
@@ -1,0 +1,9 @@
+package net.weavemc.loader.api.event
+
+public abstract class Event
+
+public abstract class CancellableEvent : Event() {
+    @get:JvmName("isCancelled")
+    public var cancelled: Boolean = false
+
+}
