+++ src/main/kotlin/net/weavemc/loader/api/event/client/ClientEvent.kt
@@ -1,0 +1,5 @@
+package net.weavemc.loader.api.event.client
+
+import net.weavemc.loader.api.event.Event
+
+public object ShutdownEvent : Event()
