+++ src/main/kotlin/net/weavemc/loader/api/event/client/GuiEvent.kt
@@ -1,0 +1,6 @@
+package net.weavemc.loader.api.event.client
+
+import net.minecraft.client.gui.GuiScreen
+import net.weavemc.loader.api.event.CancellableEvent
+
+public class GuiOpenEvent(public val screen: GuiScreen?) : CancellableEvent()
