+++ src/main/kotlin/net/weavemc/loader/hooks/ChatSentEventHook.kt
@@ -2,7 +2,7 @@ package net.weavemc.loader.hooks
 
 import net.weavemc.loader.api.Hook
 import net.weavemc.loader.api.event.CancellableEvent
-import net.weavemc.loader.api.event.ChatSentEvent
+import net.weavemc.loader.api.event.client.ChatEvent
 import net.weavemc.loader.api.util.asm
 import net.weavemc.loader.util.callEvent
 import net.weavemc.loader.util.internalNameOf
@@ -13,12 +13,12 @@ import org.objectweb.asm.tree.LabelNode
 internal class ChatSentEventHook : Hook("net/minecraft/client/entity/EntityPlayerSP") {
     override fun transform(node: ClassNode, cfg: AssemblerConfig) {
         node.methods.named("sendChatMessage").instructions.insert(asm {
-            new(internalNameOf<ChatSentEvent>())
+            new(internalNameOf<ChatEvent.Sent>())
             dup
             dup
             aload(1)
             invokespecial(
-                internalNameOf<ChatSentEvent>(),
+                internalNameOf<ChatEvent.Sent>(),
                 "<init>",
                 "(L${internalNameOf<String>()};)V"
             )
