+++ src/main/kotlin/net/weavemc/loader/hooks/KeyboardEventHook.kt
@@ -1,9 +1,11 @@
 package net.weavemc.loader.hooks
 
 import net.weavemc.loader.api.Hook
-import net.weavemc.loader.api.event.KeyboardEvent
+import net.weavemc.loader.api.event.client.KeyboardEvent
 import net.weavemc.loader.api.util.asm
-import net.weavemc.loader.util.*
+import net.weavemc.loader.util.callEvent
+import net.weavemc.loader.util.internalNameOf
+import net.weavemc.loader.util.named
 import org.objectweb.asm.tree.ClassNode
 import org.objectweb.asm.tree.MethodInsnNode
