+++ src/main/kotlin/net/weavemc/loader/hooks/ModInitializerHook.kt
@@ -2,10 +2,7 @@ package net.weavemc.loader.hooks
 
 import net.weavemc.loader.WeaveLoader
 import net.weavemc.loader.api.Hook
-import net.weavemc.loader.api.event.StartGameEvent
 import net.weavemc.loader.api.util.asm
-import net.weavemc.loader.util.callEvent
-import net.weavemc.loader.util.getSingleton
 import net.weavemc.loader.util.internalNameOf
 import net.weavemc.loader.util.named
 import org.objectweb.asm.Opcodes
