+++ src/main/kotlin/net/weavemc/loader/hooks/PlayerTickEventHook.kt
@@ -1,6 +1,7 @@
 package net.weavemc.loader.hooks
 
 import net.weavemc.loader.api.Hook
+import net.weavemc.loader.api.event.client.PlayerTickEvent
 import net.weavemc.loader.api.util.asm
 import net.weavemc.loader.util.callEvent
 import net.weavemc.loader.util.getSingleton
@@ -12,12 +13,12 @@ internal class PlayerTickEventHook : Hook("net/minecraft/entity/player/EntityPla
     override fun transform(node: ClassNode, cfg: AssemblerConfig) {
         val onUpdate = node.methods.named("onUpdate")
         onUpdate.instructions.insert(asm {
-            getSingleton<net.weavemc.loader.api.event.PlayerTickEvent.Pre>()
+            getSingleton<PlayerTickEvent.Pre>()
             callEvent()
         })
 
         onUpdate.instructions.insertBeforeReturn(asm {
-            getSingleton<net.weavemc.loader.api.event.PlayerTickEvent.Post>()
+            getSingleton<PlayerTickEvent.Post>()
             callEvent()
         })
     }
