+++ src/main/kotlin/net/weavemc/loader/hooks/WorldEventHook.kt
@@ -1,11 +1,10 @@
 package net.weavemc.loader.hooks
 
 import net.weavemc.loader.api.Hook
-import net.weavemc.loader.api.event.WorldEvent
+import net.weavemc.loader.api.event.client.WorldEvent
 import net.weavemc.loader.api.util.asm
 import net.weavemc.loader.util.callEvent
 import net.weavemc.loader.util.internalNameOf
-import jdk.internal.org.objectweb.asm.Type
 import org.objectweb.asm.tree.ClassNode
 import org.objectweb.asm.tree.LabelNode
